requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==5.1.1
Django==4.1.1
djangorestframework==3.13.1
drf-yasg==1.21.3
psycopg2==2.8.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
validators==0.20.0
yapf==0.32.0
django_redis==5.2.0